refactor: simplify ChatState and integrate useChatModel for improved state management
